digital_twin/requirements.txt
flask
ollama
pymongo
sentence-transformers
faiss-cpu
numpy==1.26.4
TTS
soundfile